+++ backend/os/location.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 
 	"github.com/c2fo/vfs/v6"
+	"github.com/c2fo/vfs/v6/options"
 	"github.com/c2fo/vfs/v6/utils"
 )
 
@@ -38,13 +39,13 @@ func (l *Location) NewFile(fileName string) (vfs.File, error) {
 
 // DeleteFile deletes the file of the given name at the location. This is meant to be a short cut for instantiating a
 // new file and calling delete on that with all the necessary error handling overhead.
-func (l *Location) DeleteFile(fileName string) error {
+func (l *Location) DeleteFile(fileName string, opts ...options.DeleteOption) error {
 	file, err := l.NewFile(fileName)
 	if err != nil {
 		return err
 	}
 
-	return file.Delete()
+	return file.Delete(opts...)
 }
 
 type fileTest func(fileName string) bool